+++ dataproxy/setup.go
@@ -50,6 +50,11 @@ func Setup(ctx context.Context, sc *app.SetupContext) error {
 	sc.Mux.Handle(clusterPath, clusterHandler)
 	logger.Infof(ctx, "Mounted ClusterService at %s", clusterPath)
 
+	translatorSvc := NewTranslatorService()
+	translatorPath, translatorHandler := workflowconnect.NewTranslatorServiceHandler(translatorSvc)
+	sc.Mux.Handle(translatorPath, translatorHandler)
+	logger.Infof(ctx, "Mounted TranslatorService at %s", translatorPath)
+
 	sc.AddReadyCheck(func(r *http.Request) error {
 		baseContainer := sc.DataStore.GetBaseContainerFQN(r.Context())
 		if baseContainer == "" {
